fix: element position indexing issue on view-transition-api calls.
When a view transition is applied to different elements, it seems like all these elements 'transition' on each and every view-transition-api call. Whenever a view-transition occures, all elements impacted by a transition come to the surface of the DOM. Their position index changes and they stack ontop of other elements that might be in a positon of absolute or fixed. This commit fixes that for the Promotion.svelte 'Claim Now' button>